requirements.txt
transformers==4.38.1
ml_collections
fschat>=0.2.36
tokenizers
huggingface_hub
numpy
sentence-transformers
tokenizers
tqdm
pandas
accelerate
protobuf
datasets
fschat
flash-attn
cmake
lit
openai==0.28.0
tenacity
google-generativeai